[#32] feature모듈안에 멀티 모듈로 작성
