--- app/src/main/java/com/mashup/kkyuni/playlist/domain/model/Date.kt
@@ -1,6 +1,0 @@
-package com.mashup.kkyuni.playlist.domain.model
-
-data class Date (
-    val year: Int,
-    val month: Int
-)
