--- app/src/main/java/com/mashup/kkyuni/playlist/presenter/holder/PlayListViewHolder.kt
@@ -1,7 +1,0 @@
-package com.mashup.kkyuni.playlist.presenter.holder
-
-import com.mashup.kkyuni.playlist.domain.model.MusicModel
-
-interface PlayListViewHolder {
-    fun bind(item: MusicModel)
-}
