+++ feature/playlist/data/src/main/java/com/mashup/kkyuni/playlist/data/repository/PlayListRepository.kt
@@ -1,0 +1,10 @@
+package com.mashup.kkyuni.playlist.data.repository
+
+import com.mashup.kkyuni.playlist.data.dto.PlayListResponseDto
+import kotlinx.coroutines.flow.Flow
+
+interface PlayListRepository {
+    fun fetchPlayList(
+        date: String
+    ): Flow<PlayListResponseDto>
+}
