+++ feature/playlist/domain/src/main/java/com/mashup/kkyuni/feature/playlist/domain/model/Date.kt
@@ -1,0 +1,6 @@
+package com.mashup.kkyuni.feature.playlist.domain.model
+
+data class Date (
+    val year: Int,
+    val month: Int
+)
