+++ feature/playlist/presentation/src/main/java/com/mashup/kkyuni/feature/playlist/presentation/PlayListBindingAdapter.kt
@@ -1,0 +1,16 @@
+package com.mashup.kkyuni.feature.playlist.presentation
+
+import androidx.databinding.BindingAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.mashup.kkyuni.feature.playlist.domain.model.PlayListState
+
+@BindingAdapter("updatePlayList")
+fun RecyclerView.updatePlayList(viewState: PlayListState){
+    (viewState as? PlayListState.Success)?.let { success ->
+        adapter?.let {
+            if(it is PlayListAdapter){
+                it.updatePlayList(success.list)
+            }
+        }
+    }
+}
