+++ feature/playlist/presentation/src/main/java/com/mashup/kkyuni/feature/playlist/presentation/holder/PlayListViewHolder.kt
@@ -1,0 +1,7 @@
+package com.mashup.kkyuni.feature.playlist.presentation.holder
+
+import com.mashup.kkyuni.feature.playlist.domain.model.MusicModel
+
+interface PlayListViewHolder {
+    fun bind(item: MusicModel)
+}
