+++ feature/playlist/presentation/src/main/java/com/mashup/kkyuni/feature/playlist/presentation/holder/impl/PlayListMusicViewHolderImpl.kt
@@ -1,0 +1,18 @@
+package com.mashup.kkyuni.feature.playlist.presentation.holder.impl
+
+import androidx.recyclerview.widget.RecyclerView
+import com.mashup.kkyuni.feature.playlist.domain.model.MusicModel
+import com.mashup.kkyuni.feature.playlist.presentation.databinding.HolderPlayListMusicBinding
+import com.mashup.kkyuni.feature.playlist.presentation.holder.PlayListViewHolder
+
+class PlayListMusicViewHolderImpl(
+    private val binding: HolderPlayListMusicBinding
+): RecyclerView.ViewHolder(binding.root), PlayListViewHolder {
+
+    override fun bind(item: MusicModel) {
+        binding.run {
+            this.item = item
+            executePendingBindings()
+        }
+    }
+}
